backend/requirements.txt
pandas
scikit-learn
fastapi
uvicorn
httpx
pydantic
requests
xgboost
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
# optional vector database client; install if you plan to use Pinecone
pinecone
